modularized create-cycle-app code
